feat: add history page
