--- lib/header.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Header extends StatelessWidget with PreferredSizeWidget {
-  @override
-  Size get preferredSize => Size.fromHeight(kToolbarHeight);
-
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      title: Text('飲む気スイッチ'),
-    );
-  }
-}
